Fixed lightning effects during 10% of expected time Not using RecycleTimer anymore but using Ticker instead A lot of work on the Ticker system
git-svn-id: svn://svn.belboudou.com/Jass/Framework@37 14830902-480a-42a1-9178-d310ea1b726b

+++ Systems/Ticker.j
@@ -15,6 +15,7 @@
 //	- tick.isExpired() : returns true or false, depending if the ticker is still ticking
 //	- local handle t = tick.getHandle() : returns handle associated to this ticker, allowing to attach data to it
 //	- tick.getRemaining() : return the number of remaining ticks
+//	- tick.getCurrent() : return the current tick
 //	- tick.reset() : reset the duration of the ticker
 //	- tick.setPeriodFactor() : set a period factor. Setting 0,5 will make ticks spawn in half the time they were intended
 //	- tick.getPeriodFactor() : retrieve the period factor
@@ -29,7 +30,7 @@ library Ticker initializer init needs Hashtable, Constants
 
 	globals
 		// Maximum TickPerSecond is 8192, timers can't ticks more often than this value
-		constant integer TickPerSecond = 1024
+		constant integer TickPerSecond = 512
 		private Tick array Ticks
 		private integer count
 		private integer tick
@@ -42,13 +43,18 @@ library Ticker initializer init needs Hashtable, Constants
 	function GetTicker takes nothing returns Tick
 		return Ticks[HTLoadInteger(GetTriggeringTrigger(), TICKER_INDEX)]
 	endfunction
+	
+	function GetTickerByTrigger takes trigger trig returns Tick
+		return Ticks[HTLoadInteger(trig, TICKER_INDEX)]
+	endfunction
 
 	struct Tick
 		// Cannot be changed
 		private trigger trig
 		private triggeraction action
 		private real period
-
+		
+		boolean infinite
 		integer last
 		integer next
 		integer done
@@ -56,6 +62,9 @@ library Ticker initializer init needs Hashtable, Constants
 		real periodFactor
 
 		// public
+		method getTrigger takes nothing returns trigger
+			return .trig
+		endmethod
 		method getHandle takes nothing returns handle
 			return .trig
 		endmethod
@@ -64,6 +73,12 @@ library Ticker initializer init needs Hashtable, Constants
 		method getRemaining takes nothing returns integer
 			return .left
 		endmethod
+		
+		// Current tick
+		method getCurrent takes nothing returns integer
+			return .done
+		endmethod
+		
 		method reset takes nothing returns nothing
 			set .left = .left + .done
 			set .done = 0
@@ -79,7 +94,15 @@ library Ticker initializer init needs Hashtable, Constants
 		method getLastTick takes nothing returns integer
 			return .last
 		endmethod
-
+		
+		// Infinite
+		method setInfinite takes boolean b returns nothing
+			set .infinite = b
+		endmethod
+		method isInfinite takes nothing returns boolean
+			return .infinite
+		endmethod
+		
 		// Period factors
 		method setPeriodFactor takes real factor returns nothing
 			set .periodFactor = factor
@@ -93,6 +116,9 @@ library Ticker initializer init needs Hashtable, Constants
 
 		method isExpired takes nothing returns boolean
 			return .left <= .last
+		endmethod
+		method isStarted takes nothing returns boolean
+			return .next > 0
 		endmethod
 
 		method start takes nothing returns nothing
@@ -121,6 +147,7 @@ library Ticker initializer init needs Hashtable, Constants
 			set .trig = CreateTrigger()
 			set .action = TriggerAddAction(.trig, func)
 			set .period = period
+			set .infinite = false
 
 			set .last = 0
 			set .next = -1
@@ -131,6 +158,7 @@ library Ticker initializer init needs Hashtable, Constants
 			call .setIndex(count)
 			set Ticks[count] = this
 			set count = count + 1
+			set Ticks[count] = 0
 			if autostart then
 				call .start()
 			endif
@@ -155,25 +183,29 @@ library Ticker initializer init needs Hashtable, Constants
 		
 		loop
 			exitwhen i >= count
+			if Ticks[i] > 0 then
+				if Ticks[i].next > -1 then
+					if Ticks[i].next <= tick then
+						set Ticks[i].left = Ticks[i].left - 1
+						set Ticks[i].done = Ticks[i].done + 1
+						
+						if Ticks[i].left < Ticks[i].last and not Ticks[i].isInfinite() then
+							call Ticks[i].delete()
 
-			if Ticks[i].next > -1 then
-				if Ticks[i].next <= tick then
-					set Ticks[i].left = Ticks[i].left - 1
-					set Ticks[i].done = Ticks[i].done + 1
-					
-					if Ticks[i].left < Ticks[i].last then
-						call Ticks[i].delete()
-
-						set count = count - 1
-						if i < count then
-							set Ticks[i] = Ticks[count]
-							call Ticks[i].setIndex(i)
+							set count = count - 1
+							if i < count then
+								set Ticks[i] = Ticks[count]
+								call Ticks[i].setIndex(i)
+								set Ticks[count] = 0
+							endif
+						else
+							call Ticks[i].exec()
+							if Ticks[i].isExpired() and not Ticks[i].isInfinite() then
+								set Ticks[i].next = tick + 1
+							else
+								call Ticks[i].nextTick()
+							endif
 						endif
-
-						set i = i - 1
-					else
-						call Ticks[i].nextTick()
-						call Ticks[i].exec()
 					endif
 				endif
 			endif
@@ -184,6 +216,7 @@ library Ticker initializer init needs Hashtable, Constants
 	private function init takes nothing returns nothing
 		set count = 0
 		set tick = 0
+		set Ticks[count] = 0
 		call TimerStart(CreateTimer(), 1. / I2R(TickPerSecond), true, function TickerActions)
 	endfunction
 endlibrary
